refactor: Dropdown 컴포넌트의 hooks 제거 및 context 격리, eslint 빈 함수 rule 수정
