+++ .eslintrc.json
@@ -6,6 +6,7 @@
     "plugin:@typescript-eslint/recommended"
   ],
   "rules": {
-    "no-console": "warn"
+    "no-console": "warn",
+    "@typescript-eslint/no-empty-function": "off"
   }
 }
