+++ src/features/common/components/dropdown/Dropdown.tsx
@@ -1,35 +1,32 @@
-import { JSXElementConstructor, ReactNode, useRef } from 'react';
+import { ReactNode, useRef, useState } from 'react';
 import Input from '@/features/common/components/dropdown/Input.tsx';
 import Title from '@/features/common/components/dropdown/Title.tsx';
 import SubmitButton from '@/features/common/components/dropdown/SubmitButton.tsx';
 import List from '@/features/common/components/dropdown/List.tsx';
 import Trigger from '@/features/common/components/dropdown/Trigger.tsx';
 import * as style from '@/features/common/components/dropdown/Dropdown.css.ts';
-import { useDropdown } from '@/features/common/hooks/useDropdown.ts';
+import DropdownContext from '@/features/common/components/dropdown/contexts/DropdownContext.ts';
+import { useClickOutSide } from '@/features/common/hooks/useClickOutSide.ts';
 
 export default function Dropdown({ children }: { children: ReactNode[] }) {
   const containerRef = useRef(null);
-  const { DropdownContext, contextValue } = useDropdown(containerRef);
+  const [isShow, setIsShow] = useState(false);
+  const [triggerRef, setTriggerRef] = useState<HTMLButtonElement | null>(null);
 
-  const trigger = children?.find(child => {
-    if (typeof child === 'object' && child !== null && 'type' in child) {
-      return (
-        (child?.type as JSXElementConstructor<unknown>)?.name === 'Trigger'
-      );
-    }
-  });
-  children = children.filter(child => {
-    if (typeof child === 'object' && child !== null && 'type' in child) {
-      return (
-        (child?.type as JSXElementConstructor<unknown>)?.name !== 'Trigger'
-      );
-    }
+  const contextValue = {
+    isShow,
+    setIsShow,
+    triggerRef,
+    setTriggerRef,
+  };
+
+  useClickOutSide(containerRef, () => {
+    setIsShow(false);
   });
 
   return (
     <DropdownContext.Provider value={contextValue}>
       <div className={style.container} ref={containerRef}>
-        {trigger}
         {children}
       </div>
     </DropdownContext.Provider>
