+++ src/features/common/components/dropdown/SubmitButton.tsx
@@ -1,6 +1,6 @@
 import * as style from '@/features/common/components/dropdown/Dropdown.css.ts';
 import { MouseEvent, MouseEventHandler, ReactNode, useContext } from 'react';
-import { DropdownContext } from '@/features/common/hooks/useDropdown.ts';
+import DropdownContext from '@/features/common/components/dropdown/contexts/DropdownContext.ts';
 
 type Props = {
   children?: ReactNode;
